+++ desktop/kde/applications/akonadi-notes/pspec.xml
@@ -1,0 +1,81 @@
+<?xml version="1.0" ?>
+<!DOCTYPE PISI SYSTEM "http://www.pisilinux.org/projeler/pisi/pisi-spec.dtd">
+<PISI>
+    <Source>
+        <Name>akonadi-notes</Name>
+        <Homepage>http://www.kde.org</Homepage>
+        <Packager>
+            <Name>Pisi Linux Admins</Name>
+            <Email>admin@pisilinux.org</Email>
+        </Packager>
+        <License>LGPLv2</License>
+        <IsA>library</IsA>
+        <IsA>app:console</IsA>
+        <Summary>Libraries and daemons to implement management of notes in Akonadi</Summary>
+        <Description>Libraries and daemons to implement management of notes in Akonadi</Description>
+        <Archive sha1sum="9cf674706757b8f09a64bcae93d152b42b064d51" type="tarxz">mirrors://kde/stable/applications/16.08.1/src/akonadi-notes-16.08.1.tar.xz</Archive>
+        <BuildDependencies>
+            <Dependency>qt5-base-devel</Dependency>
+            <Dependency>boost-devel</Dependency>
+            <Dependency>ki18n-devel</Dependency>
+            <Dependency>kdoctools-devel</Dependency>            
+            <Dependency>kitemmodels-devel</Dependency>
+            <Dependency>gettext-devel</Dependency>            
+            <Dependency>akonadi-devel</Dependency>
+            <Dependency>kdelibs4-support-devel</Dependency>
+            <Dependency>extra-cmake-modules</Dependency>            
+            <Dependency>kxmlgui-devel</Dependency>            
+            <Dependency>libxslt-devel</Dependency>
+            <Dependency>extra-cmake-modules</Dependency>
+            <Dependency>kmime-devel</Dependency>
+            <Dependency>shared-mime-info</Dependency>
+        </BuildDependencies>
+    </Source>
+
+    <Package>
+        <Name>akonadi-notes</Name>
+        <RuntimeDependencies>
+            <Dependency>qt5-base</Dependency>
+            <Dependency>ki18n</Dependency>
+            <Dependency>akonadi</Dependency>
+            <Dependency>libgcc</Dependency>
+            <Dependency>kio</Dependency>
+            <Dependency>kmime</Dependency>            
+            <Dependency>kconfig</Dependency>
+            <Dependency>kxmlgui</Dependency>            
+            <Dependency>kcoreaddons</Dependency>
+            <Dependency>kdbusaddons</Dependency>            
+            <Dependency>kwidgetsaddons</Dependency>            
+        </RuntimeDependencies>
+        <Files>
+            <Path fileType="data">/usr/share</Path>
+            <Path fileType="library">/usr/lib/qt5</Path>
+            <Path fileType="library">/usr/lib</Path>
+            <Path fileType="doc">/usr/share/doc</Path>
+        </Files>
+    </Package>
+
+    <Package>
+        <Name>akonadi-notes-devel</Name>
+        <Summary>Development files for akonadi-notes</Summary>
+        <RuntimeDependencies>
+            <Dependency>qt5-base-devel</Dependency>
+            <Dependency release="current">akonadi-notes</Dependency>
+        </RuntimeDependencies>
+        <Files>
+            <Path fileType="header">/usr/include</Path>
+            <Path fileType="data">/usr/lib/cmake</Path>
+            <Path fileType="config">/usr/lib/pkgconfig</Path>
+        </Files>
+    </Package>
+
+    <History>
+        <Update release="1">
+            <Date>2016-10-11</Date>
+            <Version>16.08.1</Version>
+            <Comment>First Release.</Comment>
+            <Name>Stefan Gronewold</Name>
+            <Email>groni@pisilinux.org</Email>
+        </Update>        
+    </History>
+</PISI>
