+++ desktop/kde/applications/kdepim-apps-libs/pspec.xml
@@ -12,7 +12,7 @@
         <IsA>library</IsA>
         <Summary>KDE PIM mail related libraries</Summary>
         <Description>KDE PIM mail related libraries</Description>
-        <Archive sha1sum="090f9cf1c7dba103924718c6cf67c57769974914" type="tarxz">mirrors://kde/stable/applications/16.04.3/src/kdepim-apps-libs-16.04.3.tar.xz</Archive>
+        <Archive sha1sum="c373a349c34be4e949efdf4ef26382130260dba4" type="tarxz">mirrors://kde/stable/applications/16.08.1/src/kdepim-apps-libs-16.08.1.tar.xz</Archive>
         <BuildDependencies>
             <Dependency>qt5-base-devel</Dependency>
             <Dependency>boost-devel</Dependency>
@@ -29,7 +29,7 @@
             <Dependency>extra-cmake-modules</Dependency>
             <Dependency>libkleo-devel</Dependency>
             <Dependency>grantlee-qt5-devel</Dependency>
-            <Dependency>pimcommon-devel</Dependency>
+            <Dependency>kdepim-common-devel</Dependency>
             <Dependency>akonadi-devel</Dependency>
             <Dependency>grantleetheme-devel</Dependency>
             <Dependency>kdelibs4-support-devel</Dependency>
@@ -38,6 +38,7 @@
             <Dependency>gpgmepp-devel</Dependency>
             <Dependency>akonadi-calendar-devel</Dependency>
             <Dependency>akonadi-search-devel</Dependency>
+            <Dependency>akonadi-contacts</Dependency>
             <Dependency>kpimtextedit-devel</Dependency>
             <Dependency>kimap-devel</Dependency>
             <Dependency versionFrom="2.4.41">openldap-client</Dependency>
@@ -74,13 +75,13 @@
             <Dependency>sonnet</Dependency>
             <Dependency>libkleo</Dependency>
             <Dependency>kdewebkit</Dependency>
-            <Dependency>pimcommon</Dependency>
-            <Dependency>kdepimlibs</Dependency>
+            <Dependency>kdepim-common</Dependency>            
             <Dependency>qt5-webkit</Dependency>
             <Dependency>grantlee-qt5</Dependency>
             <Dependency>kpimtextedit</Dependency>
             <Dependency>grantleetheme</Dependency>
             <Dependency>akonadi</Dependency>
+            <Dependency>akonadi-contacts</Dependency>
         </RuntimeDependencies>
         <Files>
             <Path fileType="config">/etc/xdg</Path>
@@ -110,6 +111,13 @@
     </Package>
 
     <History>
+        <Update release="3">
+            <Date>2016-10-11</Date>
+            <Version>16.08.1</Version>
+            <Comment>Version bump.</Comment>
+            <Name>Stefan Gronewold</Name>
+            <Email>groni@pisilinux.org</Email>
+        </Update>
         <Update release="2">
             <Date>2016-07-31</Date>
             <Version>16.04.3</Version>
