+++ desktop/kde/applications/kdepim-common/pspec.xml
@@ -1,0 +1,165 @@
+<?xml version="1.0" ?>
+<!DOCTYPE PISI SYSTEM "http://www.pisilinux.org/projeler/pisi/pisi-spec.dtd">
+<PISI>
+    <Source>
+        <Name>kdepim-common</Name>
+        <Homepage>http://www.kde.org</Homepage>
+        <Packager>
+            <Name>PisiLinux Community</Name>
+            <Email>admins@pisilinux.org</Email>
+        </Packager>
+        <License>LGPLv2</License>
+        <IsA>library</IsA>
+        <Summary>Common libraries for KDE PIM</Summary>
+        <Description>Common libraries for KDE PIM</Description>
+        <Archive sha1sum="c7a0515bb449b3c1c9d72f5dbae423eacf2a355b" type="tarxz">mirrors://kde/stable/applications/16.08.1/src/pimcommon-16.08.1.tar.xz</Archive>
+        <BuildDependencies>
+            <Dependency>qt5-base-devel</Dependency>
+            <Dependency>boost-devel</Dependency>
+            <Dependency>gpgme-devel</Dependency>
+            <Dependency>qt5-sql-sqlite</Dependency>
+            <Dependency>qt5-sql-mysql</Dependency>
+            <Dependency>qt5-sql-odbc</Dependency>
+            <Dependency>qt5-sql-postgresql</Dependency>
+            <Dependency>prison-qt5-devel</Dependency>
+            <Dependency>kmbox-devel</Dependency>
+            <Dependency>qt5-phonon-devel</Dependency>
+            <Dependency>akonadi-devel</Dependency>
+            <Dependency>libical-devel</Dependency>
+            <Dependency>libqjson-devel</Dependency>
+            <Dependency>libgpg-error-devel</Dependency>
+            <Dependency>libxslt-devel</Dependency>
+            <Dependency versionFrom="2.4.41">openldap-client</Dependency>
+            <Dependency versionFrom="2.1.26">cyrus-sasl-devel</Dependency>
+            <Dependency>kitemviews-devel</Dependency>
+            <Dependency>kio-devel</Dependency>
+            <Dependency>kdelibs4-support-devel</Dependency>
+            <Dependency>kdesignerplugin</Dependency>
+            <Dependency>kdoctools-devel</Dependency>
+            <Dependency>kemoticons-devel</Dependency>
+            <Dependency>kitemmodels-devel</Dependency>
+            <Dependency>kinit-devel</Dependency>
+            <Dependency>kunitconversion-devel</Dependency>
+            <Dependency>kmime-devel</Dependency>
+            <Dependency>kcontacts-devel</Dependency>
+            <Dependency>kcalcore-devel</Dependency>
+            <Dependency>kldap-devel</Dependency>
+            <Dependency>libkdepim-devel</Dependency>
+            <Dependency>extra-cmake-modules</Dependency>
+            <Dependency>akonadi-calendar-devel</Dependency>
+            <Dependency>akonadi-search-devel</Dependency>
+            <Dependency>qt5-webkit-devel</Dependency>
+            <Dependency>grantlee-qt5-devel</Dependency>
+            <Dependency>kdoctools-devel</Dependency>
+            <Dependency>knewstuff-devel</Dependency>
+            <Dependency>kpimtextedit-devel</Dependency>
+            <Dependency>kimap-devel</Dependency>
+        </BuildDependencies>
+    </Source>
+
+    <Package>
+        <Name>kdepim-common</Name>
+        <RuntimeDependencies>
+            <Dependency>qt5-base</Dependency>
+            <Dependency>kio</Dependency>
+            <Dependency>ki18n</Dependency>
+            <Dependency>libgcc</Dependency>
+            <Dependency>kcodecs</Dependency>
+            <Dependency>kmime</Dependency>
+            <Dependency>kldap</Dependency>
+            <Dependency>kxmlgui</Dependency>
+            <Dependency>kwallet</Dependency>
+            <Dependency>karchive</Dependency>
+            <Dependency>kjobwidgets</Dependency>
+            <Dependency>qt5-webengine</Dependency>
+            <Dependency>akonadi-contacts</Dependency>
+            <Dependency>kcalcore</Dependency>
+            <Dependency>kcontacts</Dependency>
+            <Dependency>kconfig</Dependency>
+            <Dependency>libxml2</Dependency>
+            <Dependency>kservice</Dependency>
+            <Dependency>kguiaddons</Dependency>
+            <Dependency>kitemviews</Dependency>
+            <Dependency>kcompletion</Dependency>
+            <Dependency>kcoreaddons</Dependency>
+            <Dependency>kdbusaddons</Dependency>
+            <Dependency>kiconthemes</Dependency>
+            <Dependency>kitemmodels</Dependency>
+            <Dependency>ktextwidgets</Dependency>
+            <Dependency>kconfigwidgets</Dependency>
+            <Dependency>kwidgetsaddons</Dependency>
+            <Dependency>kdelibs4-support</Dependency>
+            <Dependency>qt5-phonon</Dependency>
+            <Dependency>prison-qt5</Dependency>
+            <Dependency>akonadi</Dependency>
+            <Dependency>libkdepim</Dependency>
+            <Dependency>akonadi-calendar</Dependency>
+            <Dependency>akonadi-search</Dependency>
+            <Dependency>qt5-webkit</Dependency>
+            <Dependency>kdoctools</Dependency>
+            <Dependency>knewstuff</Dependency>
+            <Dependency>kpimtextedit</Dependency>
+            <Dependency>grantlee-qt5</Dependency>
+            <Dependency>kimap</Dependency>
+            <Dependency versionFrom="2.1.26">cyrus-sasl</Dependency>
+        </RuntimeDependencies>
+        <Files>
+            <Path fileType="config">/etc/xdg</Path>
+            <Path fileType="library">/usr/lib</Path>
+            <Path fileType="doc">/usr/share/doc</Path>
+            <Path fileType="data">/usr/mkspecs/</Path>
+            <Path fileType="executable">/usr/bin</Path>
+            <Path fileType="data">/usr/share/mime</Path>
+            <Path fileType="data">/usr/share/icons</Path>
+            <Path fileType="data">/usr/share/kservices5</Path>
+            <Path fileType="data">/usr/share/kservicetypes5</Path>
+            <Path fileType="data">/usr/share/dbus-1</Path>
+            <Path fileType="data">/usr/share/config.kcfg</Path>
+            <Path fileType="data">/usr/share/kf5</Path>
+            <Path fileType="data">/usr/share/config</Path>
+            <Path fileType="data">/usr/share/akonadi</Path>
+            <Path fileType="data">/usr/share/akonadicontact/grantleetheme/default</Path>
+        </Files>
+    </Package>
+
+    <Package>
+        <Name>kdepim-common-devel</Name>
+        <Summary>Development package for kdepim-common libraries</Summary>
+        <Description>Contains development tools and header files for kdepim-common</Description>
+        <RuntimeDependencies>
+            <Dependency>qt5-base-devel</Dependency>
+            <Dependency>boost-devel</Dependency>
+            <Dependency release="current">kdepim-common</Dependency>
+        </RuntimeDependencies>
+        <Files>
+            <Path fileType="data">/usr/lib/cmake</Path>
+            <Path fileType="header">/usr/include</Path>
+            <Path fileType="data">/usr/lib/gpgmepp</Path>
+            <Path fileType="data">/usr/share/kde4/apps/cmake</Path>
+        </Files>
+    </Package>
+
+    <History>
+        <Update release="3">
+            <Date>2016-10-11</Date>
+            <Version>16.08.1</Version>
+            <Comment>Version bump.</Comment>
+            <Name>Stefan Gronewold</Name>
+            <Email>groni@pisilinux.org</Email>
+        </Update>
+        <Update release="2">
+            <Date>2016-07-31</Date>
+            <Version>16.04.3</Version>
+            <Comment>Version bump.</Comment>
+            <Name>Yusuf Aydemir</Name>
+            <Email>yusuf.aydemir@pisilinux.org</Email>
+        </Update>
+        <Update release="1">
+            <Date>2016-04-25</Date>
+            <Version>16.04.0</Version>
+            <Comment>First release</Comment>
+            <Name>Yusuf Aydemir</Name>
+            <Email>yusuf.aydemir@pisilinux.org</Email>
+        </Update>
+    </History>
+</PISI>
