+++ desktop/kde/applications/kdepim-runtime/pspec.xml
@@ -14,7 +14,7 @@
         <IsA>app:gui</IsA>
         <Summary>KDE5 PIM Runtime Environment</Summary>
         <Description>kdepim-runtime contains KDE5 PIM (Personal Information Management) runtime environment like akonadi agents.</Description>
-        <Archive sha1sum="9d257a70c1507b8021cea9091566210ebe028e3c" type="tarxz">mirrors://kde/stable/applications/16.04.3/src/kdepim-runtime-16.04.3.tar.xz</Archive>
+        <Archive sha1sum="dca69934f85f06fd3f06b4bd1f75e233e5e4d027" type="tarxz">mirrors://kde/stable/applications/16.08.1/src/kdepim-runtime-16.08.1.tar.xz</Archive>
         <BuildDependencies>
             <Dependency>kdoctools-devel</Dependency>
             <Dependency>boost-devel</Dependency>
@@ -39,8 +39,7 @@
             <Dependency>kimap-devel</Dependency>
             <Dependency>libkgapi-devel</Dependency>
             <Dependency>kholidays-devel</Dependency>
-            <Dependency>ksyndication-devel</Dependency>
-            <Dependency>kdepimlibs-devel</Dependency>
+            <Dependency>ksyndication-devel</Dependency>            
             <Dependency>cyrus-sasl-devel</Dependency>
             <Dependency>qt5-declarative-devel</Dependency>
             <Dependency>qt5-quick1-devel</Dependency>
@@ -71,8 +70,7 @@
             <Dependency>kservice</Dependency>
             <Dependency>qt5-base</Dependency>
             <Dependency>kalarmcal</Dependency>
-            <Dependency>kcalutils</Dependency>
-            <Dependency>kdepimlibs</Dependency>
+            <Dependency>kcalutils</Dependency>            
             <Dependency>kitemviews</Dependency>
             <Dependency>kcompletion</Dependency>
             <Dependency>kcoreaddons</Dependency>
@@ -90,6 +88,11 @@
             <Dependency>qt5-xmlpatterns</Dependency>
             <Dependency>akonadi-calendar</Dependency>
             <Dependency>kdelibs4-support</Dependency>
+            <Dependency>cyrus-sasl</Dependency>
+            <Dependency>akonadi-mime</Dependency>
+            <Dependency>akonadi-notes</Dependency>
+            <Dependency>qt5-webengine</Dependency>
+            <Dependency>akonadi-contacts</Dependency>
             <Dependency>kidentitymanagement</Dependency>
         </RuntimeDependencies>
         <Files>
@@ -106,10 +109,18 @@
             <Path fileType="data">/usr/share/ontology</Path>
             <Path fileType="data">/usr/share/autostart</Path>
             <Path fileType="data">/usr/share/applications</Path>
+            <Path fileType="doc">/usr/share/doc/HTML/en/</Path>
         </Files>
     </Package>
 
     <History>
+        <Update release="4">
+            <Date>2016-10-11</Date>
+            <Version>16.08.1</Version>
+            <Comment>Version bump.</Comment>
+            <Name>Stefan Gronewold</Name>
+            <Email>groni@pisilinux.org</Email>
+        </Update>
         <Update release="3">
             <Date>2016-07-31</Date>
             <Version>16.04.3</Version>
