+++ desktop/kde/applications/kimap/pspec.xml
@@ -13,7 +13,7 @@
         <IsA>app:console</IsA>
         <Summary>Job-based API for interacting with IMAP servers</Summary>
         <Description>API for interacting with IMAP servers</Description>
-        <Archive sha1sum="83fdcec73c0c33b5c9eb698d417123eb56342a23" type="tarxz">mirrors://kde/stable/applications/16.04.3/src/kimap-16.04.3.tar.xz</Archive>
+        <Archive sha1sum="da2fa3e65239b3314955b8e58a7916f40f3282d0" type="tarxz">mirrors://kde/stable/applications/16.08.1/src/kimap-16.08.1.tar.xz</Archive>
         <BuildDependencies>
             <Dependency>qt5-base-devel</Dependency>
             <Dependency>kdoctools-devel</Dependency>
@@ -64,6 +64,13 @@
     </Package>
 
     <History>
+        <Update release="4">
+            <Date>2016-10-11</Date>
+            <Version>16.08.1</Version>
+            <Comment>Version bump.</Comment>
+            <Name>Stefan Gronewold</Name>
+            <Email>groni@pisilinux.org</Email>
+        </Update>
         <Update release="3">
             <Date>2016-07-31</Date>
             <Version>16.04.3</Version>
