+++ desktop/kde/applications/konsole/pspec.xml
@@ -11,7 +11,7 @@
         <License>GPLv2</License>
         <Summary>KDE Konsole</Summary>
         <Description>Konsole for KDE5</Description>
-        <Archive sha1sum="bc65641361c2f4d7fe0bba0109f96721c0c22a2c" type="tarxz">mirrors://kde/stable/applications/16.04.3/src/konsole-16.04.3.tar.xz</Archive>
+        <Archive sha1sum="5b98c6b08731a66561bf83248ed0951aaad93d3c" type="tarxz">mirrors://kde/stable/applications/16.08.1/src/konsole-16.08.1.tar.xz</Archive>
         <BuildDependencies>
             <Dependency>qt5-base-devel</Dependency>
             <Dependency>qt5-script-devel</Dependency>
@@ -28,6 +28,8 @@
             <Dependency>docbook-xsl</Dependency>
             <Dependency>extra-cmake-modules</Dependency>
             <Dependency>kdoctools-devel</Dependency>
+            <Dependency>kcrash-devel</Dependency>
+            <Dependency>kdbusaddons-devel</Dependency>
             <Dependency>cmake</Dependency>
         </BuildDependencies>
         <Patches>
@@ -50,6 +52,8 @@
             <Dependency>kconfig</Dependency>
             <Dependency>kxmlgui</Dependency>
             <Dependency>kservice</Dependency>
+            <Dependency>kcrash</Dependency>
+            <Dependency>kdbusaddons</Dependency>
             <Dependency>kbookmarks</Dependency>
             <Dependency>kguiaddons</Dependency>
             <Dependency>kcompletion</Dependency>
@@ -68,6 +72,13 @@
     </Package>
 
     <History>
+        <Update release="4">
+            <Date>2016-10-11</Date>
+            <Version>16.08.1</Version>
+            <Comment>Version bump.</Comment>
+            <Name>Stefan Gronewold</Name>
+            <Email>groni@pisilinux.org</Email>
+        </Update>
         <Update release="3">
             <Date>2016-07-31</Date>
             <Version>16.04.3</Version>
