+++ desktop/kde/applications/libkdepim/pspec.xml
@@ -12,7 +12,7 @@
         <IsA>library</IsA>
         <Summary>Libraries for KDE PIM applications</Summary>
         <Description>Libraries for KDE PIM applications</Description>
-        <Archive sha1sum="fbb84b5c0bf9eceb0f142b6b1785f92b649edddb" type="tarxz">mirrors://kde/stable/applications/16.04.3/src/libkdepim-16.04.3.tar.xz</Archive>
+        <Archive sha1sum="699b698934e5ff4cd4cd28067335373de729a8d3" type="tarxz">mirrors://kde/stable/applications/16.08.1/src/libkdepim-16.08.1.tar.xz</Archive>
         <BuildDependencies>
             <Dependency>qt5-base-devel</Dependency>
             <Dependency>kdoctools-devel</Dependency>
@@ -34,9 +34,14 @@
             <Dependency>kmime-devel</Dependency>
             <Dependency>kcmutils-devel</Dependency>
             <Dependency>kldap-devel</Dependency>
+            <Dependency>kio-devel</Dependency>
+            <Dependency>kcompletion-devel</Dependency>
+            <Dependency>kcoreaddons-devel</Dependency>
+            <Dependency>kconfigwidgets-devel</Dependency>
+            <Dependency>kwidgetsaddons-devel</Dependency>
             <Dependency>cyrus-sasl-devel</Dependency>
             <Dependency>kcontacts-devel</Dependency>
-            <Dependency>akonadi-calendar-devel</Dependency>
+            <Dependency>akonadi-contacts-devel</Dependency>
             <Dependency>akonadi-search-devel</Dependency>
         </BuildDependencies>
     </Source>
@@ -51,13 +56,18 @@
             <Dependency>kconfig</Dependency>
             <Dependency>kservice</Dependency>
             <Dependency>akonadi-search</Dependency>
-            <Dependency>akonadi-calendar</Dependency>
+            <Dependency>akonadi-contacts</Dependency>
             <Dependency>kjobwidgets</Dependency>
             <Dependency>kitemmodels</Dependency>
             <Dependency>kdelibs4-support</Dependency>
             <Dependency>kunitconversion</Dependency>
             <Dependency>kemoticons</Dependency>
             <Dependency>kinit</Dependency>
+            <Dependency>kio</Dependency>
+            <Dependency>kcompletion</Dependency>
+            <Dependency>kcoreaddons</Dependency>
+            <Dependency>kconfigwidgets</Dependency>
+            <Dependency>kwidgetsaddons</Dependency>
             <Dependency>solid</Dependency>
             <Dependency>kmime</Dependency>
             <Dependency>kcmutils</Dependency>
@@ -66,8 +76,7 @@
             <Dependency>kcontacts</Dependency>
             <Dependency>akonadi</Dependency>
             <Dependency>kcodecs</Dependency>
-            <Dependency>kwallet</Dependency>
-            <Dependency>kdepimlibs</Dependency>
+            <Dependency>kwallet</Dependency>            
             <Dependency>kitemviews</Dependency>
             <Dependency>kiconthemes</Dependency>
         </RuntimeDependencies>
@@ -99,6 +108,13 @@
     </Package>
 
     <History>
+        <Update release="4">
+            <Date>2016-10-11</Date>
+            <Version>16.08.1</Version>
+            <Comment>Version bump.</Comment>
+            <Name>Stefan Gronewold</Name>
+            <Email>groni@pisilinux.org</Email>
+        </Update>
         <Update release="3">
             <Date>2016-07-31</Date>
             <Version>16.04.3</Version>
