+++ desktop/kde/applications/libkleo/pspec.xml
@@ -10,9 +10,9 @@
         </Packager>
         <License>LGPLv2</License>
         <IsA>library</IsA>
-        <Summary>Common libraries for KDE PIM</Summary>
-        <Description>Common libraries for KDE PIM</Description>
-        <Archive sha1sum="4c9d5573706f6cd2a3934e41d9e8799d6dd44322" type="tarxz">mirrors://kde/stable/applications/16.04.1/src/libkleo-16.04.1.tar.xz</Archive>
+        <Summary>KDE PIM cryptographic library</Summary>
+        <Description>KDE PIM cryptographic library</Description>
+        <Archive sha1sum="80b695f0dd8e27c5fef0bad394526a44e5df3133" type="tarxz">mirrors://kde/stable/applications/16.08.1/src/libkleo-16.08.1.tar.xz</Archive>
         <BuildDependencies>
             <Dependency>qt5-base-devel</Dependency>
             <Dependency>boost-devel</Dependency>
@@ -122,6 +122,13 @@
     </Package>
 
     <History>
+        <Update release="2">
+            <Date>2016-10-11</Date>
+            <Version>16.08.1</Version>
+            <Comment>Version Bump</Comment>
+            <Name>Stefan Gronewold</Name>
+            <Email>groni@pisilinux.org</Email>
+        </Update>
         <Update release="1">
             <Date>2016-04-25</Date>
             <Version>16.04.1</Version>
