+++ pisi-index.xml.sha1sum
@@ -1,1 +1,1 @@
-183309800f642df353903621c4f90e1053a830b0
+08f2926b545d0431ca24f06559ca8d7b309ad545
