+++ pisi-index.xml.xz.sha1sum
@@ -1,1 +1,1 @@
-02ee3b829b46385362e2ab3617baaf4c0a0e75c8
+807437b9eeb304277cecdd209b1e5a8a13955e69
